Выносит render в отдельный компонент
