+++ src/helpers/utils.js
@@ -74,30 +74,4 @@ const groupEventsByDate = (events) => {
 
 const capFirstLetter = (word) => word[0].toUpperCase() + word.slice(1);
 
-const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-  return newElement.firstChild;
-};
-
-const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  AFTEREND: `afterend`,
-  BEFOREEND: `beforeend`
-};
-
-const render = (container, element, position) => {
-  switch (position) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.AFTEREND:
-      container.after(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
-export {castTimeFormat, formatTime24H, getStringDate, getISOStringDate, calcDuration, sortEventsByDate, groupEventsByDate, capFirstLetter, createElement, RenderPosition, render};
+export {castTimeFormat, formatTime24H, getStringDate, getISOStringDate, calcDuration, sortEventsByDate, groupEventsByDate, capFirstLetter};
