+++ src/main.js
@@ -1,4 +1,5 @@
-import {groupEventsByDate, sortEventsByDate, render, RenderPosition} from "./helpers/utils";
+import {groupEventsByDate, sortEventsByDate} from "./helpers/utils";
+import {render, RenderPosition} from "./helpers/render";
 import TripInfo from "./components/page-header/trip-info";
 import TripRoute from "./components/page-header/trip-route";
 import TripControls from "./components/page-header/trip-controls";
